Generalise show function
```haskell
show :: Show a => a -> Text
```

This function works great in almost any situation. But some functions accept `String` or even `Bytestring`. So in such cases `String` is converted to `Text` and then to `Bytestring` or back to `String`. Which isn't nice. And for `Text` type new implementation should work almost the same as previous implementation (as far as I understand, the only overhead - is to search for the right implementation in the dictionary, right?).